# Avoid temporary maps while reducing config