Fixes: #34

Performance test with the current config: .wotanrc.yaml -> wotan:latest -> wotan:recommended
before: 15,5 to 17,0ms
after 12,8 to 14,3ms

That's an improvement of 10 to 20% when there are only a few rules. This will be even more when the config files get bigger.